e2e-tests/requirements.txt
selenium==4.16.0
pytest==7.4.3
pytest-html==4.1.1
webdriver-manager==4.0.1
python-dotenv==1.0.0
openpyxl==3.1.2
